Actually use minified css.
The CSS file name indicated that the CSS would be minified, but in reality it's not.

I've renamed that file to `bootstrap.css`, minified it and added the minified version.
The enqueue call will now check the `wp-config` `SCRIPT_DEBUG` constant to see whether or not the minified version should be served or the development version and enqueue the correct file accordingly.

I've also rename the css file slug to `cmb2_grid_bootstrap_light` (was `bootstrap_light`) to avoid conflict with other plugins which might use the generic `bootstrap_light` slug.

Lastly, I've added a version number to the class and use that in the css file loading to allow the file to break out of the cache if the CSS would change in a future version.