Implement autoloading of the class files.
This allows for `lazy` loading, i.e. only loading those files we actually need if and when we need them.

This does presume that the SPL extension is available, but as SPL is compiled and shipped with PHP by default since PHP 5.0 and can no longer be disabled since PHP 5.3.0, this shouldn't really be an issue (except for those rare cases where self-compiled versions of PHP would be used).